# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/v1/inbox": {
      "get": {
        "operationId": "InboxController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "unreadOnly",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "id": "59eebdbd-69f5-4fc3-8f61-ac655f4812b5",
                      "staffId": "5799da62-bd38-4ddf-8171-2b2592d79c0c",
                      "type": "risk.flag_raised",
                      "severity": "HIGH",
                      "title": "Risk flag on acc_swift: volume spike",
                      "body": "3,000 sends today against a 7-day average of 214",
                      "link": "/customers/acc_swift",
                      "readAt": null,
                      "emailedAt": null,
                      "createdAt": "2026-09-24T10:48:04.229Z"
                    },
                    {
                      "id": "f6b72ba7-6ae1-484f-80cf-e09e0577d3e5",
                      "staffId": "5799da62-bd38-4ddf-8171-2b2592d79c0c",
                      "type": "senderid.sla_breach",
                      "severity": "HIGH",
                      "title": "Sender ID LEKKIPAY is past its review SLA",
                      "body": "Due 2026-09-24T06:59:59.177Z",
                      "link": "/sender-ids/d19269ff-ee3b-4fa6-b5ed-d136bad40bf1",
                      "readAt": null,
                      "emailedAt": null,
                      "createdAt": "2026-09-24T07:11:13.526Z"
                    }
                  ],
                  "hasMore": false,
                  "nextCursor": null,
                  "unread": 10
                }
              }
            }
          }
        },
        "tags": [
          "Inbox"
        ]
      }
    },
    "/admin/v1/inbox/{id}/read": {
      "post": {
        "operationId": "InboxController_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Inbox"
        ]
      }
    },
    "/admin/v1/inbox/read-all": {
      "post": {
        "operationId": "InboxController_markAllRead",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Inbox"
        ]
      }
    },
    "/admin/v1/events": {
      "get": {
        "operationId": "InboxController_events",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Live console notifications for this staff member.",
        "tags": [
          "Inbox"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zapiack Admin Console API",
    "description": "Backend for the Zapiack Admin Console.\n\n**Authentication is two steps.** `POST /auth/login` returns `totp_required` and\nsets a session cookie that every guarded route rejects until\n`POST /auth/2fa/verify` succeeds. Send `credentials: \"include\"` on every request.\n\n**Writes need `x-csrf-token`**, echoing the value from `/auth/2fa/verify` or\n`/auth/me`. Some also need an `Idempotency-Key` header and a written `reason`;\neach operation says which.\n\n**Every list returns `{ data, hasMore, nextCursor }`.** Extra context sits beside\n`data`, never instead of it.\n\n**Money is a decimal string** (`\"125000.0000\"`), never a number. Dates are\nAfrica/Lagos calendar days. Unknown query or body fields are rejected with 400.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "zpk_admin_at",
        "description": "Session cookie, set by the login flow. Not readable from JavaScript."
      }
    },
...
  }
}